requirements.txt
absl-py==0.4.1
astor==0.7.1
bleach==1.5.0
certifi==2018.11.29
chardet==3.0.4
confu==0.0.1
cpools==0.0.0
cycler==0.10.0
Cython==0.29.1
decorator==4.1.2
dlib==19.16.0
easydict==1.9
enum34==1.1.6
gast==0.2.0
graphviz==0.8.4
grpcio==1.15.0
h5py==2.7.0
html5lib==0.9999999
idna==2.7
ipython==6.1.0
ipython-genutils==0.2.0
jedi==0.10.2
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
leveldb==0.194
Markdown==2.6.11
matplotlib==1.5.0
mxnet-cu80==1.3.1
networkx==1.11
ninja-syntax==1.7.2
numpy==1.15.4
opcodes==0.3.14
opencv-python==3.4.3.18
pandas==0.20.3
PeachPy==0.2.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==5.3.0
prompt-toolkit==1.0.15
protobuf==3.6.1
ptyprocess==0.5.2
pycocotools==0.0.0
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.7.3
python-gflags==3.1.1
pytz==2018.5
PyWavelets==0.5.2
PyYAML==3.13
rawpy==0.13.0
requests==2.20.1
scikit-image==0.13.0
scikit-learn==0.20.1
scipy==0.19.1
simplegeneric==0.8.1
six==1.11.0
tensorboard==1.10.0
tensorflow-gpu==1.3.0
tensorflow-tensorboard==0.1.8
termcolor==1.1.0
traitlets==4.3.2
urllib3==1.24.1
wcwidth==0.1.7
Werkzeug==0.14.1
